enum.py renamed to subs.py. No changes to file contents.
relative import changed in __init__.py
