+++ obsws_python/__init__.py
@@ -1,6 +1,6 @@
-from .enum import Subs
 from .events import EventClient
 from .reqs import ReqClient
+from .subs import Subs
 from .version import version as __version__
 
 __ALL__ = ["ReqClient", "EventClient", "Subs"]
